// src/app/Model/AR Management/Configuration/clientlogin.ts
export class Clientlogin {
    nloginid: number;
    nclientid: number;
    nproviderid: string;
    npayerid: number;
    nupdatetoid: string;
    sloginname: string;
    spassword: string;
    sloginusing: string;
    sloginemailid: string;
    sloginemailpassword: string;
    dtloginstartdate: string;
    dtloginendate: string;
    createdon: string;
    modifiedon: string;
    bisactivelogin: boolean;
    supdatetousername: string;
}